backend/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
PyJWT
passlib[bcrypt]
sqlalchemy
pdfminer.six
python-docx
werkzeug